Fixed Logout
Last known bug is that it's not possible to logout from the user Profile page , but only directly after the instagram cookies have been added. Once the user navigates to another tab and comes back the logout functionality works on that page too. The bug is related to JSON.stringify(IG_SHORT_TOKEN). [probably]